requirements.txt
# Core API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# AI & LLM
groq>=0.4.0

# Embeddings & Vector DB
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
numpy>=1.26.0

# Parsing
pdfplumber>=0.10.4
pymupdf>=1.23.21
python-docx>=1.1.0

# NLP
spacy>=3.7.2
